Cleanup and refactor components
