+++ src/components/three-fiber/canvas-router.tsx
@@ -1,14 +1,6 @@
-import {Canvas} from '@react-three/fiber';
-import {useCallback, useEffect, useMemo, useRef, useState} from 'react';
-import {
-  getCustomDefinitions,
-  getSelectedDefinition,
-} from 'src/store/definitionsSlice';
-import {useSize} from 'src/utils/use-size';
-import {useLocation} from 'wouter';
-import {Camera} from './camera';
-import {ConfigureKeyboard, Design, Test} from '../n-links/keyboard';
-import {useAppDispatch, useAppSelector} from 'src/store/hooks';
+import {faSpinner, faUnlock} from '@fortawesome/free-solid-svg-icons';
+import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
+import {a, config, useSpring} from '@react-spring/three';
 import {
   Html,
   OrbitControls,
@@ -16,29 +8,38 @@ import {
   useGLTF,
   useProgress,
 } from '@react-three/drei';
+import {Canvas} from '@react-three/fiber';
+import {DefinitionVersionMap, KeyColorType} from '@the-via/reader';
+import cubeySrc from 'assets/models/cubey.glb';
+import glbSrc from 'assets/models/keyboard_components.glb';
+import React, {useCallback, useEffect, useMemo, useRef, useState} from 'react';
+import {shallowEqual} from 'react-redux';
 import {
+  getCustomDefinitions,
+  getSelectedDefinition,
+} from 'src/store/definitionsSlice';
+import {reloadConnectedDevices} from 'src/store/devicesThunks';
+import {useAppDispatch, useAppSelector} from 'src/store/hooks';
+import {
+  getConfigureKeyboardIsSelectable,
   getLoadProgress,
   updateSelectedKey,
-  getConfigureKeyboardIsSelectable,
 } from 'src/store/keymapSlice';
-import {a, config, useSpring} from '@react-spring/three';
-import React from 'react';
-import {shallowEqual} from 'react-redux';
-import {Object3D} from 'three';
-import {DefinitionVersionMap, KeyColorType} from '@the-via/reader';
-import {UpdateUVMaps} from './update-uv-maps';
 import {
   getDesignDefinitionVersion,
   getSelectedTheme,
 } from 'src/store/settingsSlice';
-import glbSrc from 'assets/models/keyboard_components.glb';
-import cubeySrc from 'assets/models/cubey.glb';
-import {AccentButtonLarge} from '../inputs/accent-button';
-import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
-import {reloadConnectedDevices} from 'src/store/devicesThunks';
-import {faSpinner, faUnlock} from '@fortawesome/free-solid-svg-icons';
-import {LoaderCubey} from './loader-cubey';
 import {OVERRIDE_HID_CHECK} from 'src/utils/override';
+import {useSize} from 'src/utils/use-size';
+import {Object3D} from 'three';
+import {useLocation} from 'wouter';
+import {AccentButtonLarge} from '../inputs/accent-button';
+import {ConfigureKeyboard} from '../n-links/keyboard/configure';
+import {Design} from '../n-links/keyboard/design';
+import {Test} from '../n-links/keyboard/test';
+import {Camera} from './camera';
+import {LoaderCubey} from './loader-cubey';
+import {UpdateUVMaps} from './update-uv-maps';
 useGLTF.preload(cubeySrc);
 useGLTF.preload(glbSrc);
